Refactor: 예치금 long ->int 로 변경

+++ 4Week_Misson/src/main/java/com/ll/exam/sen_books/app/cash/entity/CashLog.java
@@ -21,7 +21,7 @@ import javax.persistence.ManyToOne;
 public class CashLog extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private Member member;
-    private long price; // 변동 금액
+    private int price; // 변동 금액
     private String eventType; // 변동 사유
 
     public CashLog(long id) {
